Consolidate mcp logic inside matrix.c
