+++ keyboards/handwired/ferris/matrix.c
@@ -24,6 +24,85 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 #include QMK_KEYBOARD_H
+#include "i2c_master.h"
+
+extern i2c_status_t mcp23017_status;
+#define I2C_TIMEOUT 1000
+
+// For a better understanding of the i2c protocol, this is a good read:
+// https://www.robot-electronics.co.uk/i2c-tutorial
+
+// I2C address:
+// See the datasheet, section 3.3.1 on addressing I2C devices and figure 3-6 for an
+// illustration
+// http://ww1.microchip.com/downloads/en/devicedoc/20001952c.pdf
+// All address pins of the mcp23017 are connected to the ground on the ferris
+// | 0  | 1  | 0  | 0  | A2 | A1 | A0 |
+// | 0  | 1  | 0  | 0  | 0  | 0  | 0  |
+#define I2C_ADDR 0b0100000
+#define I2C_ADDR_WRITE ((I2C_ADDR << 1) | I2C_WRITE)
+#define I2C_ADDR_READ ((I2C_ADDR << 1) | I2C_READ)
+
+// Register addresses
+// See https://github.com/adafruit/Adafruit-MCP23017-Arduino-Library/blob/master/Adafruit_MCP23017.h
+#define IODIRA 0x00  // i/o direction register
+#define IODIRB 0x01
+#define GPPUA 0x0C  // GPIO pull-up resistor register
+#define GPPUB 0x0D
+#define GPIOA 0x12  // general purpose i/o port register (write modifies OLAT)
+#define GPIOB 0x13
+#define OLATA 0x14  // output latch register
+#define OLATB 0x15
+
+bool         i2c_initialized = 0;
+i2c_status_t mcp23017_status = I2C_ADDR;
+
+uint8_t init_mcp23017(void) {
+    print("starting init");
+    mcp23017_status = I2C_ADDR;
+
+    // I2C subsystem
+    if (i2c_initialized == 0) {
+        i2c_init();  // on pins D(1,0)
+        i2c_initialized = true;
+        wait_ms(I2C_TIMEOUT);
+    }
+
+    // set pin direction
+    // - unused  : input  : 1
+    // - input   : input  : 1
+    // - driving : output : 0
+    mcp23017_status = i2c_start(I2C_ADDR_WRITE, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+    mcp23017_status = i2c_write(IODIRA, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+    // This means: we will read all the bits on GPIOA
+    mcp23017_status = i2c_write(0b11111111, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+    // This means: we will write to the pins 0-4 on GPIOB (in select_rows)
+    mcp23017_status = i2c_write(0b11110000, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+    i2c_stop();
+
+    // set pull-up
+    // - unused  : on  : 1
+    // - input   : on  : 1
+    // - driving : off : 0
+    mcp23017_status = i2c_start(I2C_ADDR_WRITE, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+    mcp23017_status = i2c_write(GPPUA, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+    // This means: we will read all the bits on GPIOA
+    mcp23017_status = i2c_write(0b11111111, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+    // This means: we will write to the pins 0-4 on GPIOB (in select_rows)
+    mcp23017_status = i2c_write(0b11110000, I2C_TIMEOUT);
+    if (mcp23017_status) goto out;
+
+out:
+    i2c_stop();
+    return mcp23017_status;
+}
 
 /* matrix state(1:on, 0:off) */
 static matrix_row_t matrix[MATRIX_ROWS];      // debounced values
