+++ keyboards/handwired/ferris/readme.md
@@ -11,6 +11,6 @@ A split 34 keys column staggered keyboard named and decorated after the rustlang
 
 Make example for this keyboard (after setting up your build environment):
 
-    qmk flash -kb handwired/ferris -km default
+    make handwired/ferris:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
